Replace GL11 with GlStateManager, good bit of testing still needed

+++ src/main/java/mekanism/client/render/tileentity/RenderResistiveHeater.java
@@ -18,7 +18,7 @@ public class RenderResistiveHeater extends TileEntitySpecialRenderer<TileEntityR
     @Override
     public void render(TileEntityResistiveHeater tileEntity, double x, double y, double z, float partialTick, int destroyStage, float alpha) {
         GlStateManager.pushMatrix();
-        GlStateManager.translate((float) x + 0.5F, (float) y + 1.5F, (float) z + 0.5F);
+        GlStateManager.translate(x + 0.5, y + 1.5, z + 0.5);
         bindTexture(MekanismUtils.getResource(ResourceType.RENDER, "ResistiveHeater.png"));
 
         MekanismRenderer.glRotateForFacing(tileEntity);
